remove app object from stain normalization apply() call
Matlab complained about having a app object in the apply() call that was being used to apply stain nomralization via applystainnormalization(). I exported applystainnormalization() to an external file to allow it to be called without invoking the app object in the apply() call.

I also have the app add the 'Resources' directory to the Matlab path during startup to enable the applystainnormalization() function to be called directly.